part dev image texture

+++ include/GglVertexBuffer.h
@@ -4,12 +4,12 @@
 namespace mystd{
     class GglVertexBuffer{
         public:
-         GglVertexBuffer(double *positions, int32_t size);
+         GglVertexBuffer(float *positions, int32_t size);
          ~GglVertexBuffer();
          void unBindVertexBuffer();
          void bindVertexBuffer();
          uint32_t id;
          private:
-          void createVertexBuffer(double *positions, int32_t size);
+          void createVertexBuffer(float *positions, int32_t size);
     };
 }
